requirements.txt
streamlit
torch==2.3.1+cpu
torchvision==0.18.1+cpu
torchaudio==2.3.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
pillow
numpy
